Update branch with new changes since last commit
- Rebased branch onto tip of master, and updated various parts of code to take
  into accounts changes from master.

+++ src/jyut-dict/logic/search/isearchobserver.h
@@ -16,7 +16,9 @@ public:
 
     virtual void callback(const std::vector<Entry> &, bool) {}
     virtual void callback(const std::vector<SourceSentence> &, bool) {}
-    virtual void callback(const std::vector<std::pair<std::string, int>> &, bool) {}
+    virtual void callback(const std::vector<std::pair<std::string, long>> &,
+                          bool)
+    {}
 
     virtual void callback(bool, const Entry &) {}
 };
